% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ckbplotr_forest_data}
\alias{ckbplotr_forest_data}
\title{Synthetic data for demoing forest_plot}
\format{
An object of class \code{data.frame} with 42 rows and 12 columns.
}
\usage{
ckbplotr_forest_data
}
\description{
Synthetic data for demoing forest_plot
}
\keyword{datasets}
