% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ckbplotr_row_labels}
\alias{ckbplotr_row_labels}
\title{Row labels data for demoing make_shape_plot}
\format{
An object of class \code{data.frame} with 21 rows and 4 columns.
}
\usage{
ckbplotr_row_labels
}
\description{
Row labels data for demoing make_shape_plot
}
\keyword{datasets}
