% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ckbplotr_shape_data}
\alias{ckbplotr_shape_data}
\title{Synthetic data for demoing shape_plot}
\format{
An object of class \code{data.frame} with 8 rows and 5 columns.
}
\usage{
ckbplotr_shape_data
}
\description{
Synthetic data for demoing shape_plot
}
\keyword{datasets}
