% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix-panel.R
\name{fix_panel}
\alias{fix_panel}
\title{Fix panel width and height of a forest plot}
\usage{
fix_panel(plot, width = NULL, height = NULL)
}
\arguments{
\item{plot}{A plot (created by forest_plot()).}

\item{width}{Width of panels. (e.g unit(50, "mm"))}

\item{height}{Height of panels. (e.g unit(150, "mm"))}
}
\value{
A gtable object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

ckb_style(), plot_like_ckb() and shape_plot() have width and height arguments and
forest_plot() has panel.width and panel.height arguments. These use
ggh4x::force_panelsizes() to fix panel sizes.
}
\keyword{internal}
