% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_forest_plot.R
\name{fix_panel_width}
\alias{fix_panel_width}
\title{Fix panel width of a forest plot}
\usage{
fix_panel_width(plot, width)
}
\arguments{
\item{plot}{A plot (created by make_forest_plot()).}

\item{width}{Width of panels. (e.g unit(100, "mm"))}
}
\value{
A gtable object
}
\description{
\code{fix_panel_width} fixes the panel width of a forest plot
}
