% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save-figure.R
\name{ggpreview}
\alias{ggpreview}
\title{Create a plot preview and display it in the Viewer pane.}
\usage{
ggpreview(...)
}
\arguments{
\item{...}{Arguments passed to \code{ggsave} and the PNG device function.}
}
\description{
This function saves a ggplot2 plot to a temporary PNG file and then embeds it in an HTML
page, which is opened in the Viewer pane.
}
\section{Device}{
The plot is saved using \code{ggsave} with the \code{png} device, regardless of what is specified in
the call, so any arguments not used by \code{ggsave} or \code{png} are ignored.
}

