% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_forest_plot.R
\name{make_forest_data}
\alias{make_forest_data}
\title{Prepares data set for a forest plot}
\usage{
make_forest_data(
  headings = NULL,
  rows = NULL,
  cols,
  colnames = NULL,
  col.key = "key",
  col.estimate = "estimate",
  col.stderr = "stderr",
  col.lci = NULL,
  col.uci = NULL,
  col.pval = NULL,
  col.left = NULL,
  col.right = NULL,
  ci.delim = ", ",
  whiteci = NULL,
  diamond = NULL,
  bold = NULL,
  boldheadings = NULL,
  exponentiate = TRUE,
  blankrows = c(1, 1, 0, 0),
  scalepoints = FALSE,
  addtext = NULL
)
}
\arguments{
\item{headings}{A data frame that contains the headings to be used for the
rows of the plot. The data frame must contain columns 'heading1', 'heading2'
and 'heading3'. Use NA if a lower level heading is not required.}

\item{rows}{A character vector. The top level headings (heading1) of rows
to be included in the plot.}

\item{cols}{A list of data frames. These should include columns or point
estimates, and standard errors or confidence interval limits. If you
specify a headings data frame, then they must also all contain a key column
with the same name (which can be specified by col.key).}

\item{colnames}{A character vector. The titles to be used for each forest plot.
If none provided, then they will be numbered 1, 2, 3 ...}

\item{col.key}{Name of column that links the headings provided in headings
  and the results given in each data frame provided in cols.

  If headings data frame is not given, then this column will be used as headings
  for each row of the plot.

  (Default: "key")}

\item{col.estimate}{Name of column that provides point estimates.
(Default: "estimate")}

\item{col.stderr}{Name of column that provides standard errors. (Default: "stderr")}

\item{col.lci}{Name of column that provides lower limit of confidence intervals.}

\item{col.uci}{Name of column that provides upper limit of confidence intervals.}

\item{col.pval}{Currently does nothing.}

\item{col.left}{A character vector of names of columns to be printed to the left of the plot.}

\item{col.right}{A character vector of names of columns to be printed to the right of the plot.}

\item{ci.delim}{Character string to separate lower and upper limits of
confidence interval. (Default: ", ")}

\item{whiteci}{A list of character vectors. List must be the same length as cols.
Identify the rows (using the key values) for which the CI should be plotted in white. (Default: NULL)}

\item{diamond}{A list of character vectors. List must be the same length as cols.
Identify the rows (using the key values) for which the estimate and CI should be plotted using a diamond. (Default: NULL)}

\item{bold}{A list of character vectors. List must be the same length as cols.
Identify the rows (using the key values) for which text should be bold.
If a row has bold text in all cols, then heading will additionally be bold. (Default: NULL)}

\item{boldheadings}{A character vector. Identify headings (using key values)
which should additionally be bold. (Default: NULL)}

\item{exponentiate}{Exponentiate estimates (and CIs) before plotting,
use log scale on the axis, and add a line at null effect. (Default: TRUE)}

\item{blankrows}{A numeric vector of length 4 specifying the number of blank rows
after a heading1, at the end of a heading1 'section', after
a heading2, and at the end of a heading2 'section. (Default: c(1, 1, 0, 0))}

\item{scalepoints}{Should the points be scaled by inverse of the standard
error? (Default: FALSE)}

\item{addtext}{A list of data frames. List must be the same length as cols.
  Data frames should contain a col.key column, and one or more of:

  1. a column named 'text' containing character strings

  2. columns named 'het_dof', 'het_stat', and 'het_p' containing character strings

  3. columns names 'trend_stat' and 'trend_p' containing character strings

  The character strings (for 'text'), heterogeneity or trend test results will
  be plotted to the right of each forest plot below the key specified in the
  col.key column.}
}
\value{
A dataset from which the plot is generated.
}
\description{
\code{make_forest_data}
}
\keyword{internal}
