% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_shape_plot.R
\name{make_shape_plot}
\alias{make_shape_plot}
\title{Make a shape plot with ggplot2}
\usage{
make_shape_plot(data, col.x = "x", col.estimate = "estimate",
  col.stderr = "stderr", col.lci = NULL, col.uci = NULL,
  col.n = NULL, col.group = NULL, exponentiate = FALSE,
  scalepoints = FALSE, pointsize = 3, xlims, ylims, gap = c(0.025,
  0.025), ext = c(0.025, 0.025), ratio = 1.5, xbreaks = NULL,
  ybreaks = NULL, xlab = "Risk factor", ylab = "Estimate (95\% CI)",
  title = "Figure", printplot = TRUE, showcode = TRUE)
}
\arguments{
\item{data}{The data frame containing estimates to be plotted.}

\item{col.x}{Name of column that provides the x-axis value (e.g. exposure, risk factor, dependent variable). (Default: "x")}

\item{col.estimate}{Name of column that provides point estimates.
(Default: "estimate")}

\item{col.stderr}{Name of column that provides standard errors. (Default: "stderr")}

\item{col.lci}{Name of column that provides lower limit of confidence intervals.}

\item{col.uci}{Name of column that provides upper limit of confidence intervals.}

\item{col.n}{Name of column that provides number to be plotted below CIs.}

\item{col.group}{Name of column that groups the estimates. (Default: NULL)}

\item{exponentiate}{Exponentiate estimates (and CIs) before plotting,
use log scale on the axis. (Default: TRUE)}

\item{scalepoints}{Should the points be scaled by inverse of the standard
error? (Default: FALSE)}

\item{xlims}{A numeric vector of length two. The limits of the x-axis.}

\item{ylims}{A numeric vector of length two. The limits of the y-axis.}

\item{gap}{A numeric vector of length two. The gap between plotting area and axis to the left and bottom of the plot, as a proportion of the x-axis length. (Default: c(0.025, 0.025))}

\item{ext}{A numeric vector of length two. The extensions to add to the right and top of the plot, as a proportion of the x-axis length. (Default: c(0.025, 0.025))}

\item{ratio}{The ratio (y-axis:x-axis) to use for the plot. (Default: 1.5)}

\item{xlab}{Label for x-axis. (Default: "Risk factor")}

\item{ylab}{Label for y-axis. (Default: "Estimate (95\% CI)")}

\item{title}{Plot title. (Default: "Figure")}

\item{printplot}{Print the plot. (Default: TRUE)}

\item{showcode}{Show the ggplot2 code to generate the plot in RStudio 'Viewer' pane. (Default: TRUE)}
}
\value{
A list:
\describe{
  \item{plot}{the plot}
  \item{code}{ggplot2 code to generate the plot}
}
}
\description{
Make a shape plot with ggplot2
}
