% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape-plot.R
\name{shape_plot}
\alias{shape_plot}
\title{Make a shape plot with ggplot2}
\usage{
shape_plot(
  data,
  col.x = "x",
  col.estimate = c("estimate", "est", "beta", "loghr"),
  col.stderr = c("stderr", "std.error", "std.err", "se"),
  col.lci = NULL,
  col.uci = NULL,
  col.n = NULL,
  exponentiate = FALSE,
  logscale = exponentiate,
  scalepoints = FALSE,
  digits = 2,
  minse = NA,
  pointsize = 3,
  col.group = NULL,
  shape = 15,
  plotcolour = "black",
  colour = plotcolour,
  cicolour = colour,
  fill = colour,
  ciunder = NULL,
  lines = FALSE,
  xlims = NULL,
  ylims = NULL,
  height = NULL,
  width = NULL,
  gap = c(0.025, 0.025),
  ext = c(0.025, 0.025),
  ratio = 1.5,
  plot.margin = margin(0.5, 1.5, 0.5, 0.5, "lines"),
  clip = "off",
  base_size = 11,
  base_line_size = base_size/22,
  stroke = base_size/22,
  axis.title.margin = 1,
  xbreaks = NULL,
  ybreaks = NULL,
  xlab = "Risk factor",
  ylab = "Estimate (95\% CI)",
  legend.name = "",
  legend.position = "top",
  title = NULL,
  quiet = FALSE,
  printplot = !quiet,
  showcode = !quiet,
  addaes = NULL,
  addarg = NULL,
  add = NULL,
  envir = NULL
)
}
\arguments{
\item{data}{The data frame containing estimates to be plotted.}

\item{col.x}{Name of column that provides the x-axis value (e.g. exposure, risk factor, dependent variable). (Default: "x")}

\item{col.estimate}{Name of column that provides point estimates.
(Default: "estimate")}

\item{col.stderr}{Name of column that provides standard errors. (Default: "stderr")}

\item{col.lci}{Name of column that provides lower limit of confidence intervals.}

\item{col.uci}{Name of column that provides upper limit of confidence intervals.}

\item{col.n}{Name of column that provides number to be plotted below CIs.}

\item{exponentiate}{Exponentiate estimates (and CIs) before plotting,
use log scale on the axis. (Default: FALSE)}

\item{logscale}{Use log scale for vertical axis. (Default: exponentiate)}

\item{scalepoints}{Should the points be scaled by inverse of the standard
error? (Default: FALSE)}

\item{digits}{Number of digits to use in text of estimates.}

\item{minse}{Minimum standard error to use when scaling point size. (Default will use minimum in the data.)}

\item{pointsize}{The (largest) size of box to use for plotting point
estimates. (Default: 3)}

\item{col.group}{Name of column that groups the estimates. (Default: NULL)}

\item{shape}{Shape of points. An integer, or name of a column of integers. (Default: 15)}

\item{plotcolour}{Colour for non-data aspects of the plot. (Default: "black")}

\item{colour}{Colour of points. Name of a colour, or name of a column of colour names. (Default will use plotcolour)}

\item{cicolour}{Colour of CI lines. Colour of CI lines. Name of a colour, or name of a column of colour names. (Default will use plotcolour)}

\item{fill}{Fill colour of points. Fill colour of points. Name of a colour, or name of a column of colour names. (Default will use plotcolour)}

\item{ciunder}{Plot CI lines before points. A logical value, or name of a column of logical values. (Default will plot CI lines after points.)}

\item{lines}{Plot lines (linear fit through estimates, weighted by inverse variance). (Default: FALSE)}

\item{xlims}{A numeric vector of length two. The limits of the x-axis.}

\item{ylims}{A numeric vector of length two. The limits of the y-axis.}

\item{height}{Panel height to use and apply different formatting to short CIs. A grid::unit() object, or if numeric is assumed to be in mm.}

\item{width}{Panel width.A grid::unit() object, or if numeric is assumed to be in mm.}

\item{gap}{A numeric vector of length two. The gap between plotting area and axis to the left and bottom of the plot, as a proportion of the x-axis length. (Default: c(0.025, 0.025))}

\item{ext}{A numeric vector of length two. The extensions to add to the right and top of the plot, as a proportion of the x-axis length. (Default: c(0.025, 0.025))}

\item{ratio}{The ratio (y-axis:x-axis) to use for the plot. (Default: 1.5)}

\item{plot.margin}{Plot margin, given as margin(top, right, bottom, left, units). (Default: margin(0.5, 1.5, 0.5, 0.5, "lines"))}

\item{clip}{Passed to clip argument of \code{ggplot2::coord_cartesian()}. (Default: "on")}

\item{base_size}{base font size, given in pts.}

\item{base_line_size}{base size for line elements}

\item{stroke}{Size of outline of shapes. (Default: base_size/22)}

\item{axis.title.margin}{Margin between axis titles and plot. (Default: 1)}

\item{xbreaks}{Breaks for the x axis. Passed to ggplots::scale_x_continuous. (Default: NULL)}

\item{ybreaks}{Breaks for the y axis. Passed to ggplots::scale_y_continuous. (Default: NULL)}

\item{xlab}{Label for x-axis. (Default: "Risk factor")}

\item{ylab}{Label for y-axis. (Default: "Estimate (95\% CI)")}

\item{legend.name}{The name of the colour scale/legend for groups. (Default: "")}

\item{legend.position}{Position of the legend for groups ("none", "left", "right", "bottom", "top", or two-element numeric vector). (Default: "top")}

\item{title}{Plot title. (Default: NULL)}

\item{quiet}{Set to TRUE to not print the plot nor show generated code in the RStudio 'Viewer' pane. (Default: FALSE)}

\item{printplot}{Print the plot. (Default: !quiet)}

\item{showcode}{Show the ggplot2 code to generate the plot in RStudio 'Viewer' pane. (Default: !quiet)}

\item{addaes, addarg, add}{Methods for customising the plot. See documentation for details.}

\item{envir}{Environment in which to evaluate the plot code. May be useful when calling this function inside another function.}
}
\value{
A list:
\describe{
\item{plot}{the plot}
\item{code}{ggplot2 code to generate the plot}
}
}
\description{
Make a shape plot with ggplot2
}
