% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_like_ckb.R
\name{ckb_style}
\alias{ckb_style}
\title{Make a ggplot into CKB style}
\usage{
ckb_style(
  xlims = NULL,
  ylims = NULL,
  gap = c(0.025, 0.025),
  ext = c(0, 0),
  ratio = 1.5,
  width = NULL,
  height = NULL,
  base_size = 11,
  base_line_size = base_size/22,
  colour = "black",
  plot.margin = margin(0.5, 1.5, 0.5, 0.5, "lines"),
  axes = "both"
)
}
\arguments{
\item{xlims}{A numeric vector of length two. The limits of the x-axis.}

\item{ylims}{A numeric vector of length two. The limits of the y-axis.}

\item{gap}{A numeric vector of length two. The gap between plotting area and axis to the left and bottom of the plot, as a proportion of the x-axis length. (Default: c(0.025, 0.025))}

\item{ext}{A numeric vector of length two. The extensions to add to the right and top of the plot, as a proportion of the x-axis length. (Default: c(0, 0))}

\item{ratio}{The ratio (y-axis:x-axis) to use for the plot. Ignored if both width and height are set. (Default: 1.5)}

\item{width}{A \code{grid::unit} object to set the width of the plot (not including the gap or extension).}

\item{height}{A \code{grid::unit} object to set the height of the plot (not including the gap or extension).}

\item{base_size}{base font size, given in pts.}

\item{base_line_size}{base size for line elements}

\item{colour}{Colour for non-data aspects of the plot. (Default: "black")}

\item{plot.margin}{Margin around entire plot (Default: margin(0.5, 0, 0.5, 0, "lines"))}

\item{axes}{Choice of axis lines to add to the plot, one of "both", "x" or "y". (Default: "both")}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
