% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ckbplotr-package.R
\docType{package}
\name{ckbplotr-package}
\alias{ckbplotr}
\alias{ckbplotr-package}
\title{ckbplotr: Create CKB Plots}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

ckbplotr provides functions to help create and style plots in R. It is being developed by, and primarily for, China Kadoorie Biobank researchers.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://neilstats.github.io/ckbplotr/}
  \item \doi{10.5281/zenodo.6382217}
  \item Report bugs at \url{https://github.com/neilstats/ckbplotr/issues}
}

}
\author{
\strong{Maintainer}: Neil Wright \email{neil.wright@ndph.ox.ac.uk}

}
\keyword{internal}
