% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix-panel.R
\name{fix_panel}
\alias{fix_panel}
\title{Fix panel width and height of a forest plot}
\usage{
fix_panel(plot, width = NULL, height = NULL)
}
\arguments{
\item{plot}{A plot (created by forest_plot()).}

\item{width}{Width of panels. (e.g unit(50, "mm"))}

\item{height}{Height of panels. (e.g unit(150, "mm"))}
}
\value{
A gtable object
}
\description{
\code{fix_panel} fixes the panel width and height of a forest plot
}
