% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_forest_plot.R
\name{make_forest_plot}
\alias{make_forest_plot}
\title{Make forest plot with ggplot2}
\usage{
make_forest_plot(headings = NULL, rows = NULL, cols,
  exponentiate = TRUE, colnames = NULL, col.key = "key",
  col.estimate = "estimate", col.stderr = "stderr", col.lci = NULL,
  col.uci = NULL, col.left = NULL, col.right = NULL,
  col.left.heading = "", col.left.space = 0, col.right.space = 0,
  col.pval = NULL, ci.delim = ", ", title = "",
  xlab = "HR (95\% CI)", xlim = NULL, xticks = NULL,
  blankrows = c(1, 1, 0, 0), whiteci = NULL, diamond = NULL,
  scalepoints = FALSE, pointsize = 3, addtext = NULL,
  heading.space = 4, plot.space = 8)
}
\arguments{
\item{headings}{A data frame that contains the headings to be used for the
rows of the plot. The data frame must contain columns 'heading1', 'heading2'
and 'heading3'. Use NA if a lower level heading is not required.}

\item{rows}{A character vector. The top level headings (heading1) of rows
to be included in the plot.}

\item{cols}{A list of data frames. These should include columns or point
estimates, and standard errors or confidence interval limits. If you
specify a headings data frame, then they must also all contain a key column
with the same name (which can be specified by col.key).}

\item{exponentiate}{Exponentiate estimates (and CIs) before plotting,
use log scale on the axis, and add a line at null effect. (Default: TRUE)}

\item{colnames}{A character vector. The titles to be used for each forest plot.
If none provided, then they will be numbered 1, 2, 3 ...}

\item{col.key}{Name of column that links the headings provided in headings
  and the results given in each data frame provided in cols.

  If headings data frame is not given, then this column will be used as headings
  for each row of the plot.

  (Default: "key")}

\item{col.estimate}{Name of column that provides point estimates.
(Default: "estimate")}

\item{col.stderr}{Name of column that provides standard errors. (Default: "stderr")}

\item{col.lci}{Name of column that provides lower limit of confidence intervals.}

\item{col.uci}{Name of column that provides upper limit of confidence intervals.}

\item{col.left}{A character vector of names of columns to be printed to the left of the plot.}

\item{col.right}{Name of additional column to be printed to the right of the plot.
Currently this is not added to the plot, but will be in the data frame.}

\item{col.left.heading}{A character vector of titles for col.left columns. (Default: "")}

\item{col.left.space}{A numeric vector. Sizes of the gaps between the plot
and col.left columns. As a multiple of the length of the x-axis. (Default: 0)}

\item{col.right.space}{Size of the gap between the plot and column to the
right of the plot. As a multiple of the length of the x-axis. (Default: 0)}

\item{col.pval}{Currently does nothing.}

\item{ci.delim}{Character string to separate lower and upper limits of
confidence interval. (Default: ", ")}

\item{title}{Title to appear at the top of the plot.}

\item{xlab}{Label to appear below the x-axis and above the text showing
estimates and CIs. (Default: "HR (95\% CI)")}

\item{xlim}{A numeric vector. The limits of the x axis.}

\item{xticks}{A numeric vector. The tick points of the x axis.}

\item{blankrows}{A numeric vector of length 4 specifying the number of blank rows
after a heading1, at the end of a heading1 'section', after
a heading2, and at the end of a heading2 'section. (Default: c(1, 1, 0, 0))}

\item{whiteci}{A list of character vectors. List must be the same length as cols.
Identify the rows (using the key values) for which the CI should be plotted in white. (Default: NULL)}

\item{diamond}{A list of character vectors. List must be the same length as cols.
Identify the rows (using the key values) for which the estimate and CI should be plotted using a diamond. (Default: NULL)}

\item{scalepoints}{Should the points be scaled by inverse of the standard
error? (Default: FALSE)}

\item{pointsize}{The (largest) size of box to use for plotting point
estimates. (Default: 3)}

\item{addtext}{A list of data frames. List must be the same length as cols.
  Data frames should contain a col.key column, and one or more of:

  1. a column named 'text' containing character strings

  2. columns named 'het_dof', 'het_stat', and 'het_p' containing character strings

  3. columns names 'trend_stat' and 'trend_p' containing character strings

  The character strings (for 'text'), heterogeneity or trend test results will
  be plotted to the right of each forest plot below the key specified in the
  col.key column.}

\item{heading.space}{Size of the gap between headings and the first plot.
Unit is "lines". (Default: 4)}

\item{plot.space}{Size of the gap between forest plots.
Unit is "lines". (Default: 8)}
}
\value{
A list:
\describe{
  \item{plot}{the plot}
  \item{data}{a data frame from which the plot is generated}
  \item{code}{ggplot2 code to generate the plot}
}
}
\description{
\code{make_forest_plot} creates a forest plot with ggplot
}
\details{
The function returns the plot, a dataset which is used to create the plot,
and the ggplot2 code that creates the plot. In RStudio, the ggplot2 code
will be shown in the viewer.
}
