% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_shape_plot.R
\name{make_shape_plot}
\alias{make_shape_plot}
\title{Make a shape plot with ggplot2}
\usage{
make_shape_plot(
  data,
  col.x = "x",
  col.estimate = "estimate",
  col.stderr = "stderr",
  col.lci = NULL,
  col.uci = NULL,
  col.n = NULL,
  exponentiate = FALSE,
  logscale = exponentiate,
  scalepoints = FALSE,
  minse = NA,
  pointsize = 3,
  col.group = NULL,
  shape = NULL,
  colour = NULL,
  cicolour = colour,
  fill = NULL,
  ciunder = NULL,
  lines = FALSE,
  xlims,
  ylims,
  gap = c(0.025, 0.025),
  ext = c(0.025, 0.025),
  ratio = 1.5,
  base_size = 11,
  base_line_size = base_size/22,
  stroke = base_size/22,
  xbreaks = NULL,
  ybreaks = NULL,
  xlab = "Risk factor",
  ylab = "Estimate (95\% CI)",
  title = "Figure",
  printplot = TRUE,
  showcode = TRUE,
  addcode = NULL,
  addaes = NULL,
  addarg = NULL
)
}
\arguments{
\item{data}{The data frame containing estimates to be plotted.}

\item{col.x}{Name of column that provides the x-axis value (e.g. exposure, risk factor, dependent variable). (Default: "x")}

\item{col.estimate}{Name of column that provides point estimates.
(Default: "estimate")}

\item{col.stderr}{Name of column that provides standard errors. (Default: "stderr")}

\item{col.lci}{Name of column that provides lower limit of confidence intervals.}

\item{col.uci}{Name of column that provides upper limit of confidence intervals.}

\item{col.n}{Name of column that provides number to be plotted below CIs.}

\item{exponentiate}{Exponentiate estimates (and CIs) before plotting,
use log scale on the axis. (Default: FALSE)}

\item{logscale}{Use log scale for vertical axis. (Default: exponentiate)}

\item{scalepoints}{Should the points be scaled by inverse of the standard
error? (Default: FALSE)}

\item{minse}{Minimum standard error to use when scaling point size. (Default will use minimum in the data.)}

\item{pointsize}{The (largest) size of box to use for plotting point
estimates. (Default: 3)}

\item{col.group}{Name of column that groups the estimates. (Default: NULL)}

\item{shape}{Shape of points. An integer, or name of a column of integers. (Default will use shape 22 - squares with fill.)}

\item{colour}{Colour of points. Name of a colour, or name of a column of colour names. (Default will use black.)}

\item{cicolour}{Colour of CI lines. Colour of CI lines. Name of a colour, or name of a column of colour names. (Default will use black.)}

\item{fill}{Fill colour of points. Fill colour of points. Name of a colour, or name of a column of colour names. (Default will use black.)}

\item{ciunder}{Plot CI lines before points. A logical value, or name of a column of logical values. (Default will plot CI lines after points.)}

\item{lines}{Plot lines (linear fit through estimates, weighted by inverse variance). (Default: FALSE)}

\item{xlims}{A numeric vector of length two. The limits of the x-axis.}

\item{ylims}{A numeric vector of length two. The limits of the y-axis.}

\item{gap}{A numeric vector of length two. The gap between plotting area and axis to the left and bottom of the plot, as a proportion of the x-axis length. (Default: c(0.025, 0.025))}

\item{ext}{A numeric vector of length two. The extensions to add to the right and top of the plot, as a proportion of the x-axis length. (Default: c(0.025, 0.025))}

\item{ratio}{The ratio (y-axis:x-axis) to use for the plot. (Default: 1.5)}

\item{base_size}{base font size, given in pts.}

\item{base_line_size}{base size for line elements}

\item{stroke}{Size of outline of shapes. (Default: base_size/22)}

\item{xbreaks}{Breaks for the x axis. Passed to ggplots::scale_x_continuous. (Default: NULL)}

\item{ybreaks}{Breaks for the y axis. Passed to ggplots::scale_y_continuous. (Default: NULL)}

\item{xlab}{Label for x-axis. (Default: "Risk factor")}

\item{ylab}{Label for y-axis. (Default: "Estimate (95\% CI)")}

\item{title}{Plot title. (Default: "Figure")}

\item{printplot}{Print the plot. (Default: TRUE)}

\item{showcode}{Show the ggplot2 code to generate the plot in RStudio 'Viewer' pane. (Default: TRUE)}

\item{addcode}{A character vector of code to add to the generated code.
The first element should be a regular expression.
The remaining elements are added to the generated code just before the first match of a line (trimmed of  whitespace) with the regular expression. (Default: NULL)}

\item{addaes}{Specify additional aesthetics for some ggplot layers.}

\item{addarg}{Specify additional arguments for some ggplot layers.}
}
\value{
A list:
\describe{
  \item{plot}{the plot}
  \item{code}{ggplot2 code to generate the plot}
}
}
\description{
Make a shape plot with ggplot2
}
