% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ckb-style.R
\name{plot_like_ckb}
\alias{plot_like_ckb}
\title{Make a ggplot into CKB style}
\usage{
plot_like_ckb(plot, ...)
}
\arguments{
\item{plot}{A ggplot2 plot}

\item{...}{Arguments passed to ckb_style()}
}
\value{
A ggplot2 plot.
}
\description{
Make a ggplot into CKB style
}
