% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save-figure.R
\name{save_figure}
\alias{save_figure}
\title{Output plots as files}
\usage{
save_figure(
  figure,
  name,
  title = "",
  titlepos = grid::unit.c(unit(1.27/2, "cm"), unit(1, "npc") - unit(1.27/2, "cm")),
  titlejust = c(0, 1),
  margin = unit(c(2.27, 1.27, 1.27, 1.27), units = "cm"),
  size = NULL,
  pagesize = c("A4", "A5"),
  landscape = FALSE,
  pagedim = NULL,
  cropped = FALSE
)
}
\arguments{
\item{figure}{Plot (or graphical object).}

\item{name}{Name of figure. Used to name the output file(s) and embedded in the PDF document properties Title field.}

\item{title}{Title to be added to the page. (Default: "")}

\item{titlepos}{Position of the title text. Default is 1/4 inch from top left of page.
(Default: unit.c(unit(1.27/2, "cm"), unit(1, "npc") - unit(1.27/2, "cm")))}

\item{titlejust}{Justification of the title text. (Default: c(0, 1))}

\item{margin}{Margin to be placed around the plot.
Default is 2.27cm top, 1.27cm (1/2 inch) other sides.
(Default: unit(c(2.27, 1.27, 1.27, 1.27), units = "cm"))}

\item{size}{A unit vector of length two (width, height).
Size of plot (a width/height larger than page weight/height minus margins will be
ignored), centred within margins.
By default, plot will fill the space within margins.}

\item{pagesize}{Page size of PDF output: "A4" or "A5". (Default: "A4")}

\item{landscape}{Landscape page orientation? (Default: False)}

\item{pagedim}{Dimensions (width, height) of PDF output. Overrides pagesize and landscape arguments if used.}

\item{cropped}{Create a PNG output of the figure without margins or title. (Default: False)}
}
\description{
Output plots as files
}
