% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ckb-style.R
\name{theme_ckb}
\alias{theme_ckb}
\title{CKB ggplot theme}
\usage{
theme_ckb(
  base_size = 11,
  base_line_size = base_size/22,
  ink = "black",
  paper = "white",
  colour = NULL,
  axis.title.margin = 1,
  plot.margin = margin(0.5, 1.5, 0.5, 0.5, "lines"),
  ...
)
}
\arguments{
\item{base_size}{Base font size, given in pts. (Default: 11)}

\item{base_line_size}{Base size for line elements. (Deault: base_size/22)}

\item{ink, paper}{Colour for foreground and background elements. (Defaults: "black" and "white")}

\item{colour}{Deprecated. Use \code{ink} instead.}

\item{axis.title.margin}{Margin between axis titles and plot. (Default: 1)}

\item{plot.margin}{Margin around entire plot (Default: margin(0.5, 0, 0.5, 0, "lines"))}

\item{...}{Arguments passed to \code{ggplot2::theme_bw()} if using ggplot2 4.0.0 or later.}
}
\description{
A ggplot2 theme, based on theme_bw.
}
