% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_like_ckb.R
\name{theme_ckb}
\alias{theme_ckb}
\title{CKB ggplot theme}
\usage{
theme_ckb(
  base_size = 11,
  base_line_size = base_size/22,
  colour = "black",
  plot.margin = margin(0.5, 0, 0.5, 0, "lines")
)
}
\arguments{
\item{base_size}{base font size, given in pts.}

\item{base_line_size}{base size for line elements}

\item{colour}{Colour for non-data aspects of the plot. (Default: "black")}

\item{plot.margin}{Margin around entire plot (Default: margin(0.5, 0, 0.5, 0, "lines"))}
}
\description{
Based on theme_bw
}
